requirements.txt
opencv-python==4.8.1.78
numpy>=1.21.0
matplotlib>=3.5.0
scikit-image>=0.19.0
scipy>=1.7.0
Pillow>=9.0.0,<10.0.0
streamlit>=1.25.0
